refactor away duplicate code
