+++ packages/prop-house-webapp/src/components/HouseManager/RoundInfoConfig/index.tsx
@@ -3,10 +3,8 @@ import Footer from '../Footer';
 import RoundFields from '../RoundFields';
 import { useDispatch } from 'react-redux';
 import { useAppSelector } from '../../../hooks';
-import { checkStepCriteria, updateRound } from '../../../state/slices/round';
-import { capitalize } from '../../../utils/capitalize';
-import removeTags from '../../../utils/removeTags';
-import { useEffect, useRef, useState } from 'react';
+import { updateRound } from '../../../state/slices/round';
+import { useEffect, useRef } from 'react';
 import { PropHouseWrapper } from '@nouns/prop-house-wrapper';
 
 const RoundInfoConfig = () => {
@@ -16,49 +14,11 @@ const RoundInfoConfig = () => {
   const host = useAppSelector(state => state.configuration.backendHost);
   const client = useRef(new PropHouseWrapper(host));
 
-  const [title, setTitle] = useState(round.title || '');
-  const [description, setDescription] = useState(round.description || '');
-  const [descriptionLength, setDescriptionLength] = useState(0);
-  const [errors, setErrors] = useState<{ title?: string; description?: string }>({});
-
-  const handleBlur = (field: 'title' | 'description') => {
-    const value = field === 'title' ? title : description;
-    const length = value === description ? descriptionLength : value.length;
-    const minLen = field === 'title' ? 5 : 20;
-    const maxLen = field === 'title' ? 255 : undefined;
-
-    const error =
-      length < minLen
-        ? `${capitalize(field)} must be at least ${minLen} characters.`
-        : maxLen && length > maxLen
-        ? `${capitalize(field)} must be less than ${maxLen} characters.`
-        : undefined;
-
-    setErrors({ ...errors, [field]: error });
-  };
-
-  const handleFieldChange = (field: 'title' | 'description', value: string) => {
-    errors[field] && setErrors({ ...errors, [field]: undefined });
-
-    field === 'title' ? setTitle(value) : setDescription(value);
-
-    dispatch(updateRound({ ...round, [field]: value }));
-    dispatch(checkStepCriteria());
-  };
-  useEffect(() => {
-    setDescriptionLength(removeTags(round.description).length);
-  }, [round.description]);
-
-  const handleDescriptionChange = (value: string) => {
-    setDescription(value);
-    handleFieldChange('description', value);
-  };
-
   useEffect(() => {
     if (round.house.existingHouse || round.house.contractURI !== '') {
       return;
     } else {
-      // upload contracturi to ipfs if the user is creating
+      // upload contractURI to ipfs if the user is creating
       // a new house or if they have not already set one
       postContractURIToIpfs();
     }
@@ -85,10 +45,7 @@ const RoundInfoConfig = () => {
       dispatch(
         updateRound({
           ...round,
-          house: {
-            ...round.house,
-            contractURI: `ipfs://${res.data.ipfsHash}`,
-          },
+          house: { ...round.house, contractURI: `ipfs://${res.data.ipfsHash}` },
         }),
       );
     } catch (e) {
@@ -99,15 +56,7 @@ const RoundInfoConfig = () => {
   return (
     <>
       <Header title="What's your round about?" />
-      <RoundFields
-        title={title}
-        description={description}
-        descriptionLength={descriptionLength}
-        errors={errors}
-        handleBlur={handleBlur}
-        handleChange={handleFieldChange}
-        handleDescriptionChange={handleDescriptionChange}
-      />
+      <RoundFields round={round} />
       <Footer />
     </>
   );
